fix: entity and service package

+++ src/main/java/com/TTecnologia/EmissorNfeTorres/dao/ClienteDao.java
@@ -4,6 +4,9 @@ import com.TTecnologia.EmissorNfeTorres.model.entity.Cliente;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.util.Optional;
+
 @Repository
-public interface ClienteDao extends JpaRepository<Cliente, Long> {
+public interface ClienteDao extends JpaRepository<Cliente, Integer> {
+
 }
